+++ CHANGES.rst
@@ -77,3 +77,6 @@ v1.0.6
 
 v1.0.7[unreleased]
 -----------
+* Add default organization type
+* Add hidden_address field to Organization
+* Use GoogleAddressSerializer instead of GoogleAddressCityStateSerializer on OrganizationRetrieveSerializer